GestionAutenticacion/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
ecdsa==0.17.0
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
